Added permission checking to the Article Routes

+++ app/Http/Controllers/ArticleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Article;
+use App\UserRole;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -14,10 +15,14 @@ class ArticleController extends Controller
     }
 
     public function create() {
+        $perm = UserRole::CREATE_ARTICLE;
+        if (!Auth::user()->role->$$perm) { return redirect('home'); }
         return view('pages.article.create');
     }
 
     public function store(Request $request) {
+        $perm = UserRole::CREATE_ARTICLE;
+        if (!Auth::user()->role->$$perm) { return redirect('home'); }
         $article = Article::create([
             'title' => $request->input('title'),
             'body' => $request->input('body'),
@@ -31,19 +36,37 @@ class ArticleController extends Controller
     }
 
     public function edit(Article $article) {
-        return view('pages.article.edit', ['article' => $article]);
+        $permOther = UserRole::MANAGE_OTHER_ARTICLE;
+        $permOwn = UserRole::MANAGE_OWN_ARTICLE;
+        $role = Auth::user()->role;
+        if ($role->$permOther || ($role->$permOwn && $article->author == Auth::id())) {
+            return view('pages.article.edit', ['article' => $article]);
+        }
+        return redirect('home');
     }
 
     public function update(Request $request, Article $article) {
-        $article->update([
-            'title' => $request->input('title'),
-            'body' => $request->input('body'),
-            'user_id' => Auth::user()->id,
-        ]);
-        return redirect()->route('view_article', ['article' => $article]);
+        $permOther = UserRole::MANAGE_OTHER_ARTICLE;
+        $permOwn = UserRole::MANAGE_OWN_ARTICLE;
+        $role = Auth::user()->role;
+        if ($role->$permOther || ($role->$permOwn && $article->author == Auth::id())) {
+            $article->update([
+                'title' => $request->input('title'),
+                'body' => $request->input('body'),
+                'user_id' => Auth::user()->id,
+            ]);
+            return redirect()->route('view_article', ['article' => $article]);
+        }
+        return redirect('home');
     }
 
     public function destroy(Article $article) {
-        $article->delete();
+        $permOther = UserRole::MANAGE_OTHER_ARTICLE;
+        $permOwn = UserRole::MANAGE_OWN_ARTICLE;
+        $role = Auth::user()->role;
+        if ($role->$permOther || ($role->$permOwn && $article->author == Auth::id())) {
+            $article->delete();
+        }
+        return redirect('home');
     }
 }
